requirements.txt
beautifulsoup4==4.13.4
requests==2.32.4
python-dotenv==1.1.1
schedule==1.2.2

fastapi==0.116.0
uvicorn==0.35.0

pydantic==2.11.7
pydantic_core==2.33.2
typing_extensions==4.14.1

starlette==0.46.2
anyio==4.9.0
sniffio==1.3.1
h11==0.16.0
